api(container-runtime): expose versionMarkResolver on IContainerRuntime
Move the version mark resolver public types (IVersionMarkResolver,
ResolveResult, VersionMarkCapture) into container-runtime-definitions and
expose `versionMarkResolver` on the `@legacy @beta` IContainerRuntime
interface, so hosts obtain the resolver from a supported interface rather
than the concrete `@internal` ContainerRuntime class. container-runtime
re-exports the types for back-compat; its API surface is unchanged.
