+++ packages/runtime/container-runtime-definitions/api-report/container-runtime-definitions.legacy.beta.api.md
@@ -25,6 +25,7 @@ export interface IContainerRuntime extends IProvideFluidDataStoreRegistry, ICont
     readonly scope: FluidObject;
     // (undocumented)
     readonly storage: IContainerStorageService;
+    readonly versionMarkResolver: IVersionMarkResolver;
 }
 
 // @beta @sealed @legacy (undocumented)
@@ -94,6 +95,24 @@ export interface ISummarizerObservabilityProps {
     numUnsummarizedRuntimeOps: number;
 }
 
+// @beta @legacy
+export interface IVersionMarkResolver {
+    onBatchSequenced(listener: (batchId: string, sequenceNumber: number, timestamp?: number) => void): () => void;
+    resolve(batchId: string, sequenceNumberLowerBound: number): Promise<ResolveResult>;
+    sealAndCaptureVersionMark(): VersionMarkCapture;
+}
+
+// @beta @legacy
+export type ResolveResult = {
+    readonly kind: "resolved";
+    readonly sequenceNumber: number;
+    readonly timestamp?: number;
+} | {
+    readonly kind: "pending";
+} | {
+    readonly kind: "unresolvable";
+};
+
 // @beta @sealed @legacy (undocumented)
 export type SummarizerStopReason =
 /**
@@ -128,6 +147,17 @@ export type SummarizerStopReason =
 * first submitSummary attempt fails for any reason and there's a 2nd summary attempt without an ack
 */
 | "latestSummaryStateStale";
+
+// @beta @legacy
+export type VersionMarkCapture = {
+    readonly kind: "pending";
+    readonly batchId: string;
+    readonly sequenceNumberLowerBound: number;
+} | {
+    readonly kind: "resolved";
+    readonly sequenceNumber: number;
+    readonly timestamp?: number;
+};
 
 // (No @packageDocumentation comment for this package)
