+++ packages/runtime/container-runtime-definitions/src/index.ts
@@ -41,3 +41,8 @@ export type {
 	ISummarizerObservabilityProps,
 	ISummarizerEvents,
 } from "./containerRuntime.js";
+export type {
+	IVersionMarkResolver,
+	ResolveResult,
+	VersionMarkCapture,
+} from "./versionMarks.js";
