+++ packages/runtime/container-runtime/src/versionMarks/index.ts
@@ -3,11 +3,13 @@
  * Licensed under the MIT License.
  */
 
+export type {
+	IVersionMarkResolver,
+	ResolveResult,
+	VersionMarkCapture,
+} from "@fluidframework/container-runtime-definitions/internal";
 export {
 	VersionMarkResolver,
-	type IVersionMarkResolver,
-	type ResolveResult,
-	type VersionMarkCapture,
 	type VersionMarkResolverRuntimeHooks,
 } from "./versionMarkResolver.js";
 export { inboundVersionMarkUpdate, type InboundVersionMarkUpdate } from "./inboundBatch.js";
